requirements.txt
fastapi
uvicorn
playwright
sqlalchemy
apscheduler
pydantic
python-dotenv
requests
psycopg2-binary
